refactoring ga.replace_duplicates submodule
